Changed rating to int type in database schema.

+++ src/main/resources/sql/init_database.sql
@@ -20,7 +20,7 @@ CREATE TABLE IF NOT EXISTS WINE (
     name TEXT UNIQUE,
     year INTEGER,
     variety TEXT,
-    rating NUMERIC,
+    rating INTEGER,
     price NUMERIC,
     colour TEXT,
     vineyard TEXT REFERENCES VINEYARD(name),
